Created ViewOnceInformation repo
ViewOnceInformation is used by views that needs to display something for the 1st time one. After the user dismisses it, we persist the fact it has been shown once. The enum can be used to add several items. Check is done simply by retrieving the array and see if it includes the desired item, OR a convenience function:

 * infoHasBeenShown: (info: InformationViewOnce) => boolean;

will return true/false by passing the emum you want to check.