fix: pending bridge txs should have a different key
Fixes a race condition bug where an incoming history transaction could have the
same hash as a pending bridge transaction (which is removed soon after). This
caused FlashList to add extra rows that never get removed from the list.

The header key was also changed to not rely on index position for the key.